refactor(slot):made some changes on interviewer slot generating process. refactor(admin:interviewerVerification):attach reason while sending rejection mail
