+++ src/controllers/interviewer/IInterviewerController.ts
@@ -4,7 +4,11 @@ export interface IInterviewerController{
     getInterviewerProfile(request: Request, response: Response): Promise<void>;
     updateInterviewerProfile(request: Request, response: Response): Promise<void>;
     changePassword(request: Request, response: Response): Promise<void>;
-    getSlotsByInterviewerId(request: Request, response: Response): Promise<void>;
-    generateSlots(request: Request, response: Response): Promise<void>;
+    createSlotGenerationRule(request: Request, response: Response): Promise<void>;
+    getSlotsByRule(
+    request: Request,
+    response: Response
+  ): Promise<void>
+  getInterviewerSlotGenerationRule(request: Request, response: Response): Promise<void>
     
 }
