+++ src/services/slot/ISlotService.ts
@@ -1,7 +1,11 @@
 import { IInterviewSlot } from "../../models/slot/interviewSlot"
 import { ISlotGenerationRule } from "../../models/slot/slotGenerationRule"
 export interface ISlotService {
-    createRuleAndGenerateSlots(ruleData:ISlotGenerationRule):Promise<IInterviewSlot[]>
+    createSlotGenerationRule(ruleData:ISlotGenerationRule):Promise<IInterviewSlot[]>
     generateSlotsFromRule(rule: ISlotGenerationRule): Promise<IInterviewSlot[]>
-    getSlotsByInterviewerId(interviewerId: string): Promise<IInterviewSlot[]|[]>
+    // getSlotsByInterviewerId(interviewerId: string): Promise<IInterviewSlot[]|[]>
+     getSlotsByRule(interviewerId: string): Promise<IInterviewSlot[] | []>
+      getInterviewerSlotGenerationRule(
+    interviewerId: string
+  ): Promise<ISlotGenerationRule | null>
 }
